+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 ----------
 - undo add setDeleted($deleted) in ConversationReceiver because already in there.
 - added Form\DeleteConversationsForm test
+- Service\PmService  test fully implemented